:truck: Refactored saving logic into mixin
